+++ app/src/main/java/ee/taltech/likutt/iti0213_2019s_hw02/TrackingLocation.kt
@@ -6,7 +6,7 @@ class TrackingLocation(var id: Int,
                        var recordedAt: String,
                        var sessionId: Long,
                        var type: String,
-                       var speed: Long?) {
+                       var speed: Double?) {
 
     override fun toString(): String {
         return "id: $id, latitude: $latitude, longitude: $longitude, recordedAt: $recordedAt, " +
